requirements.txt
certifi==2019.9.11
coverage==4.5.4
Django==2.2
django-cors-headers==3.1.1
django-material-admin==1.6.37
djangorestframework==3.10.3
entrypoints==0.3
flake8==3.7.9
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.3
sqlparse==0.3.0
numpy==1.16.1
